github-agent/github_daily_report/pyproject.toml
[project]
name = "github-daily-report"
version = "0.1.0"
description = ""
authors = [
    {name = "Ashwin"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain",
    "langchain_openai",
    "langchain-anthropic",
    "requests",
    "langgraph",
    "python-dotenv"
]

packages = [
    {include = "github_daily_report"}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
